+++ test/sources/merlin_hide.t/hidden.ml
@@ -1,0 +1,5 @@
+include struct
+  let g x = x
+
+  let f x = Mylib.truc (g x)
+end [@@merlin.hide]