+++ test/sources/merlin_hide.t/mylib.ml
@@ -1,0 +1,1 @@
+let truc x = x